refactor(dal): inline cypher generation
