+++ kg_maintainer/__init__.py
@@ -17,10 +17,6 @@ from .merge import (
     merge_character_profile_updates,
     merge_world_item_updates,
 )
-from .cypher_generation import (
-    generate_character_node_cypher,
-    generate_world_element_node_cypher,
-)
 
 __all__ = [
     "AgentStateData",
@@ -34,6 +30,4 @@ __all__ = [
     "parse_unified_world_updates",
     "merge_character_profile_updates",
     "merge_world_item_updates",
-    "generate_character_node_cypher",
-    "generate_world_element_node_cypher",
 ]
